+++ src/pages/dashboard/consultation/[uuid-consultation].tsx
@@ -1,4 +1,4 @@
-import React, { memo, ReactElement, useEffect, useState } from "react";
+import React, { memo, ReactElement, useEffect, useRef, useState } from "react";
 import { GetStaticPaths, GetStaticProps } from "next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { pdfjs } from "react-pdf";
@@ -70,6 +70,8 @@ import { appLockSelector } from "@features/appLock";
 import moment from "moment";
 import { Session } from "next-auth";
 import { DefaultCountry } from "@app/constants";
+import { useLeavePageConfirm } from "@app/hooks/useLeavePageConfirm";
+import dynamic from "next/dynamic";
 
 pdfjs.GlobalWorkerOptions.workerSrc = `//cdnjs.cloudflare.com/ajax/libs/pdf.js/${pdfjs.version}/pdf.worker.js`;
 
@@ -95,7 +97,16 @@ function ConsultationInProgress() {
   const theme = useTheme();
   const router = useRouter();
   const { data: session } = useSession();
-
+  useLeavePageConfirm(() => {
+    if (!leaveDialog.current) {
+      /*if (!window.confirm(`message: ${uuind}`)) {
+                throw "Route Canceled";
+            } else {
+                // localStorage.removeItem(`consultation-data-${uuind}`);
+            }*/
+    }
+  });
+  const leaveDialog = useRef(false);
   const [filterdrawer, setFilterDrawer] = useState(false);
   const [value, setValue] = useState<string>("consultation_form");
   const [acts, setActs] = useState<any>("");
@@ -120,7 +131,6 @@ function ConsultationInProgress() {
   const [selectedUuid, setSelectedUuid] = useState<string[]>([]);
   const [pendingDocuments, setPendingDocuments] = useState<any[]>([]);
   const [isViewerOpen, setIsViewerOpen] = useState<string>("");
-  const [sheet, setSheet] = useState<any>(null);
   const [actions, setActions] = useState<boolean>(false);
   const [loading, setLoading] = useState<boolean>(true);
   const [instruction, setInstruction] = useState("");
@@ -241,6 +251,7 @@ function ConsultationInProgress() {
         }
       : null
   );
+
   const { data: httpModelResponse } = useRequest(
     medical_entity
       ? {
@@ -281,6 +292,7 @@ function ConsultationInProgress() {
         }
       : null
   );
+
   const { data: httpSheetResponse } = useRequest(
     mpUuid && agenda
       ? {
@@ -293,6 +305,7 @@ function ConsultationInProgress() {
         }
       : null
   );
+
   const { data: httpDocumentResponse, mutate: mutateDoc } = useRequest(
     mpUuid && agenda
       ? {
@@ -344,53 +357,6 @@ function ConsultationInProgress() {
     setInfo(null);
     setOpenDialog(true);
   }, [selectedDialog, setInfo, setOpenDialog]);
-
-  useEffect(() => {
-    if (httpSheetResponse) {
-      setSheet((httpSheetResponse as HttpResponse)?.data);
-    }
-  }, [httpSheetResponse]);
-
-  useEffect(() => {
-    if (sheet) {
-      setSelectedModel(sheet.modal);
-      if (!localStorage.getItem("Modeldata" + uuind)) {
-        localStorage.setItem(
-          "Modeldata" + uuind,
-          JSON.stringify(sheet.modal.data)
-        );
-      }
-      const app_data = sheet.exam.appointment_data;
-      // load observation data from local storage
-      const examData = localStorage.getItem(`consultation-data-${uuind}`);
-      dispatch(
-        SetExam({
-          motif:
-            examData && JSON.parse(examData).motif
-              ? JSON.parse(examData).motif
-              : app_data?.consultation_reason
-              ? app_data?.consultation_reason.uuid
-              : "",
-          notes:
-            examData && JSON.parse(examData).notes
-              ? JSON.parse(examData).notes
-              : app_data?.notes
-              ? app_data.notes.value
-              : "",
-          diagnosis: app_data?.diagnostics
-            ? app_data.diagnostics.value
-            : examData
-            ? JSON.parse(examData).diagnosis
-            : "",
-          treatment: app_data?.treatments
-            ? app_data.treatments.value
-            : examData
-            ? JSON.parse(examData).treatment
-            : "",
-        })
-      );
-    }
-  }, [dispatch, sheet, uuind]);
 
   useEffect(() => {
     if (appointement) {
@@ -472,7 +438,7 @@ function ConsultationInProgress() {
   useEffect(() => {
     const acts: { act_uuid: any; name: string; qte: any; price: any }[] = [];
     if (end) {
-      if (appointement?.status !== 5) {
+      if ([5, 4].includes(appointement?.status)) {
         selectedAct.map(
           (act: { uuid: any; act: { name: string }; qte: any; fees: any }) => {
             acts.push({
@@ -492,7 +458,7 @@ function ConsultationInProgress() {
         );
         form.append("notes", exam.notes);
         form.append("diagnostic", exam.diagnosis);
-        form.append("treatment", exam.treatment);
+        form.append("treatment", exam.treatment ? exam.treatment : "");
         form.append("consultation_reason", exam.motif);
         form.append("fees", total.toString());
         form.append("consultation_fees", consultationFees.toString());
@@ -507,14 +473,16 @@ function ConsultationInProgress() {
           },
         }).then(() => {
           console.log("end consultation");
-          dispatch(setTimer({ isActive: false }));
+          appointement?.status !== 5 && dispatch(setTimer({ isActive: false }));
           mutate().then(() => {
-            localStorage.removeItem("Modeldata" + uuind);
-            localStorage.removeItem(`consultation-data-${uuind}`);
-            localStorage.removeItem(`consultation-fees`);
-            localStorage.removeItem(`consultation-acts`);
+            leaveDialog.current = true;
             router.push("/dashboard/agenda").then(() => {
+              localStorage.removeItem("Modeldata" + uuind);
+              localStorage.removeItem(`consultation-data-${uuind}`);
+              localStorage.removeItem(`consultation-fees`);
+              localStorage.removeItem(`consultation-acts`);
               setActions(false);
+              setEnd(false);
             });
             sendNotification();
           });
@@ -522,12 +490,27 @@ function ConsultationInProgress() {
       } else {
         router.push("/dashboard/agenda").then(() => {
           setActions(false);
+          setEnd(false);
         });
       }
     }
-    setEnd(false);
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [end]);
+
+  const sheet = (httpSheetResponse as HttpResponse)?.data;
+  const sheetExam = sheet?.exam;
+
+  useEffect(() => {
+    if (sheet) {
+      setSelectedModel(sheet.modal);
+      if (!localStorage.getItem("Modeldata" + uuind)) {
+        localStorage.setItem(
+          "Modeldata" + uuind,
+          JSON.stringify(sheet.modal.data)
+        );
+      }
+    }
+  }, [dispatch, sheet, uuind]);
 
   const sendNotification = () => {
     if (secretary.length > 0) {
@@ -896,21 +879,19 @@ function ConsultationInProgress() {
               )}
             </Grid>
             <Grid item xs={12} md={7} style={{ paddingLeft: 10 }}>
-              {sheet && (
-                <ConsultationDetailCard
-                  {...{
-                    changes,
-                    setChanges,
-                    uuind,
-                    agenda: agenda?.uuid,
-                    mutateDoc,
-                    medical_entity,
-                    session,
-                    router,
-                  }}
-                  exam={sheet.exam}
-                />
-              )}
+              <ConsultationDetailCard
+                {...{
+                  changes,
+                  setChanges,
+                  uuind,
+                  agenda: agenda?.uuid,
+                  exam: sheetExam,
+                  mutateDoc,
+                  medical_entity,
+                  session,
+                  router,
+                }}
+              />
             </Grid>
           </Grid>
         </TabPanel>
@@ -1019,7 +1000,7 @@ function ConsultationInProgress() {
                 )}
                 <Button
                   onClick={
-                    appointement?.status == 5
+                    appointement?.status === 5
                       ? saveConsultation
                       : endConsultation
                   }